emotion-recognition/requirements.txt
Flask
numpy
tensorflow==2.0
Pillow
flask_cors
gunicorn